Add files via upload
Environmental monitoring plays a pivotal role in assessing and mitigating the impact of human activities on our planet. This paper provides an overview of the recent advances in environmental monitoring, encompassing a range of tools, technologies, and their diverse applications. We discuss the importance of real-time data acquisition, remote sensing, and data analysis techniques in tracking air and water quality, biodiversity, climate change, and more. Furthermore, the paper emphasizes the significance of collaborative efforts among governments, scientists, and the private sector in developing sustainable solutions to address pressing environmental challenges.